+++ tests_golden/integration/test_cookiecutter_generation/api-proprietary/frequenz-api-test/docs/_scripts/mkdocstrings_autoapi.py
@@ -1,0 +1,9 @@
+# License: Proprietary
+# Copyright © 2023 Frequenz Energy-as-a-Service GmbH
+
+"""Generate the code reference pages."""
+
+from frequenz.repo.config.mkdocs import api_pages
+
+api_pages.generate_python_api_pages("py", "python-reference")
+api_pages.generate_protobuf_api_pages()