+++ tests_golden/integration/test_cookiecutter_generation/model-proprietary/frequenz-model-test/noxfile.py
@@ -1,0 +1,8 @@
+# License: Proprietary
+# Copyright © 2023 Frequenz Energy-as-a-Service GmbH
+
+"""Configuration file for nox."""
+
+from frequenz.repo.config import RepositoryType, nox
+
+nox.configure(RepositoryType.MODEL)